Makefile bugs fixed.  mpi/gui compile and work completely on i686 and x86_64 machines

+++ Makefile
@@ -12,33 +12,37 @@ MODLFLAGS=-loadflags "$(shell pwd)/libresample-0.1.3/libresample.a"
 all: clean-all libresample gui mpi
 
 clean-mpi:
-	-rm -rf mpi 
-	-rm -rf $(ARCH)
+	-[ -d mpi ] && rm -rf mpi 
 
-clean: clean-mpi
-	-rm -rf $(ARCH)
+clean-gui:
+	-[ -d gui ] && rm -rf gui
+
+clean: clean-mpi clean-gui 
+	-[ -d $(ARCH) ] && rm -rf $(ARCH)
 	-rm -f *.ras *.hist *.connect *.curvs
 	-rm -f out.dat *.[0-9]* *~ *.bak *.sav 
 
-clean-all: clean
-	-rm -rf mpi gui 
-	-(cd libresample-0.1.3; make clean)
+clean-libresample:
+	make -C libresample-0.1.3 clean
 
 
-gui:  # libresample-0.1.3/libresample.a
+clean-all: clean clean-libresample
+
+
+gui:  clean-gui libresample-0.1.3/libresample.a
 	$(GUIMODL) $(MODLFLAGS)
-#	[ -d gui ] && rm -rf gui
 	mv $(ARCH) gui
-	(cd gui; sed -i 's#cnstellate/$(ARCH)#cnstellate/gui#g' special)
+	sed -i 's#cnstellate/$(ARCH)#cnstellate/gui#g' gui/special
 
-mpi: #libresample-0.1.3/libresample.a
+mpi: clean-mpi libresample-0.1.3/libresample.a
 	$(MPIMODL) $(MODLFLAGS)
-#	[ -d mpi ] && rm -rf mpi
 	mv $(ARCH) mpi
-	(cd mpi; find -type f -exec sed -i 's#cnstellate/$(ARCH)#cnstellate/mpi#g' {} \; )
+	sed -i 's#cnstellate/$(ARCH)#cnstellate/mpi#g' mpi/special
+
 
 libresample-0.1.3/libresample.a:
-	(cd libresample-0.1.3; make CFLAGS="-fPIC $CFLAGS -fPIC" CXXFLAGS="-fPIC $CXXFLAGS -fPIC")
+	(cd libresample-0.1.3; make CFLAGS="-fPIC ${CFLAGS} -fPIC" CXXFLAGS="-fPIC ${CXXFLAGS} -fPIC" libresample.a)
+
 
 libresample_0.1.3.tar.gz:
 	wget -c http://ftp.debian.org/debian/pool/main/libr/libresample/libresample_0.1.3.orig.tar.gz
